+++ components/composer/CHANGELOG.md
@@ -8,3 +8,4 @@
 
 - [Composer] Don't hide `from` field when account is loaded and add collapsible height to whole composer [#247](https://github.com/nylas/components/pull/247)
 - [Composer] Add background color to contact search dropdown items and increase search input width [#251](https://github.com/nylas/components/pull/251)
+- [Composer] Fix `show_cc` and `show_bcc` buttons and sets defaults to `true` [#250](https://github.com/nylas/components/pull/250)